@babiabeo 2025-05-18
I'm also not in favor of this change. Using a string to represent a single char seems a bit odd to me. 🤔